Component refactoring
refactoring components structure

more :
convert price to int64
fixing search
