--- client/components/Product/RelatedRow.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const  RelatedRow = () =>  {
-  return (
-    <div>RelatedRow</div>
-  )
-}
-
-export default RelatedRow
