+++ client/components/Search/SearchPanel.js
