+++ client/components/Swipers/LatestSwiper.js
@@ -1,29 +1,27 @@
 import { Swiper, SwiperSlide } from "swiper/react";
 import { useEffect, useState } from "react";
-import Image from "next/image";
-import Link from "next/link";
 import "swiper/css";
 import { Navigation, Pagination, Autoplay, EffectFade } from "swiper";
 import "swiper/css/navigation";
 import "swiper/css/effect-fade";
 import "swiper/css/pagination";
 
-import ProductCard from "../ui/ProductCard"
-import { Box, Stack, Typography } from "@mui/material";
+import ProductCard from "../Product/ProductCard"
 import DviderTitle from "../ui/DviderTitle";
+import { Box, Stack, Typography } from "@mui/material";
 
 
-const PopularRow = ({ products: allProducts }) => {
-    const [products, setProducts] = useState([])
+const LatestSwiper = ({products: allProducts}) => {
+    const [products,setProducts] = useState([])
     useEffect(() => {
         setProducts(allProducts)
-    }, [])
+    },[])
 
     return (
         <>
             <Box>
-                <DviderTitle title="محبوب ترین" my={'20px'} />
-                <Swiper
+                <DviderTitle title="تازه ترین ها" my={'20px'}/>
+                <Swiper 
                     modules={[Navigation, Pagination, EffectFade]}
                     navigation
                     spaceBetween={2}
@@ -48,4 +46,6 @@ const PopularRow = ({ products: allProducts }) => {
     )
 }
 
-export default PopularRow
+
+
+export default LatestSwiper
