+++ client/components/Swipers/OfferSwiper.js
@@ -1,19 +1,23 @@
 import { Swiper, SwiperSlide } from "swiper/react";
 import { useEffect, useState } from "react";
-import Image from "next/image";
-import Link from "next/link";
+
 import "swiper/css";
 import { Navigation, Pagination, Autoplay, EffectFade } from "swiper";
 import "swiper/css/navigation";
 import "swiper/css/effect-fade";
 import "swiper/css/pagination";
 
-import ProductCard from "../ui/ProductCard"
-import { Box, Stack, Typography } from "@mui/material";
-import DviderTitle from "../ui/DviderTitle";
+import ProductCard from "../Product/ProductCard"
+import { Box, Typography } from "@mui/material";
+import styled from "@emotion/styled";
 
 
-const LatestRow = ({products: allProducts}) => {
+const StyledBox = styled(Box)(({theme}) => ({
+        backgroundColor : theme.palette.danger.main,
+        padding : "20px 15px"
+}))
+
+const OfferSwiper = ({products: allProducts}) => {
     const [products,setProducts] = useState([])
     useEffect(() => {
         setProducts(allProducts)
@@ -21,21 +25,22 @@ const LatestRow = ({products: allProducts}) => {
 
     return (
         <>
-            <Box>
-                <DviderTitle title="تازه ترین ها" my={'20px'}/>
-                <Swiper 
+            <StyledBox>
+                <Typography variant="h4" color="white" my={1}>پیشنهاد های امروز</Typography>
+                <Swiper
                     modules={[Navigation, Pagination, EffectFade]}
                     navigation
                     spaceBetween={2}
                     slidesPerView="auto"
+                    // centeredSlides={true}
                     loop={false}
-
+                    
                 >
                     {products.map((product, index) => (
                         <SwiperSlide
                             key={index}
                             style={{
-                                width: "auto",
+                                width : "auto",
                             }}
                         >
                             <ProductCard product={product}/>
@@ -43,11 +48,10 @@ const LatestRow = ({products: allProducts}) => {
                     ))}
 
                 </Swiper>
-            </Box>
+            </StyledBox>
         </>
     )
 }
 
+export default OfferSwiper
 
-
-export default LatestRow
