+++ client/components/Swipers/RelatedSwiper.js
@@ -1,47 +1,39 @@
 import { Swiper, SwiperSlide } from "swiper/react";
 import { useEffect, useState } from "react";
-import Image from "next/image";
-import Link from "next/link";
 import "swiper/css";
-import { Navigation, Pagination, Autoplay, EffectFade } from "swiper";
+import { Navigation, Pagination, EffectFade } from "swiper";
 import "swiper/css/navigation";
 import "swiper/css/effect-fade";
 import "swiper/css/pagination";
 
-import ProductCard from "../ui/ProductCard"
-import { Box, Typography } from "@mui/material";
-import styled from "@emotion/styled";
+import ProductCard from "../Product/ProductCard"
+import { Box } from "@mui/material";
 import DviderTitle from "../ui/DviderTitle";
 
-const StyledBox = styled(Box)(({theme}) => ({
-        backgroundColor : theme.palette.danger.main,
-        padding : "20px 15px"
-}))
 
-const OffRow = ({products: allProducts}) => {
-    const [products,setProducts] = useState([])
+const RelatedSwiper = ({ products: allProducts }) => {
+    const [products, setProducts] = useState([])
     useEffect(() => {
         setProducts(allProducts)
-    },[])
+    }, [])
 
     return (
         <>
-            <StyledBox>
-                <Typography variant="h4" color="white" my={1}>پیشنهاد های امروز</Typography>
+            <Box>
+                <DviderTitle title="محصولات مشابه" my={'20px'} />
                 <Swiper
                     modules={[Navigation, Pagination, EffectFade]}
                     navigation
                     spaceBetween={2}
                     slidesPerView="auto"
-                    // centeredSlides={true}
                     loop={false}
-                    
+
                 >
                     {products.map((product, index) => (
                         <SwiperSlide
                             key={index}
                             style={{
-                                width : "auto",
+                                width: "auto",
                             }}
                         >
                             <ProductCard product={product}/>
@@ -49,10 +41,11 @@ const OffRow = ({products: allProducts}) => {
                     ))}
 
                 </Swiper>
-            </StyledBox>
+            </Box>
         </>
     )
 }
 
-export default OffRow
 
+
+export default RelatedSwiper
