+++ client/components/Swipers/ThumbnailSwipre.js
