+++ client/pages/admin/dashboard/components/sign-in-page.js
@@ -1,6 +1,6 @@
 
 import AdminLayout from "../../../../components/Admin/AdminLayout";
-import MainLayout from "../../../../components/Layouts/MainLayout";
+import MainLayout from "../../../../layouts/MainLayout";
 
 function signInPage() {
   return (
