+++ client/pages/admin/dashboard/index.js
@@ -1,4 +1,4 @@
-import MainLayout from "../../../components/Layouts/MainLayout";
+import MainLayout from "../../../layouts/MainLayout";
 import Meta from "../../../components/Meta";
 import AdminLayout from "../../../components/Admin/AdminLayout";
 import { Box } from "@mui/material";
