+++ server/controllers/productController.js
@@ -1,6 +1,7 @@
 const Category = require("../models/categories")
 const Brand = require("../models/brands")
 const Product = require("../models/products")
+const _ = require("lodash")
 
 exports.getAllBrands = async (req, res, next) => {
     try {
@@ -52,7 +53,7 @@ exports.getAllProducts = async (req, res, next) => {
         const searching = req.query.search || req.query.categories || discount || price ||brand ? true : false;
         const itemPerPage = req.query.limit
         const sortBy = req.query.sort || false
-
+        
         //mongoDb Queries 
         const placeHolderQ = {'none' : ""}
         const textSearchQuery = search ? { $text: { $search: search } } : placeHolderQ
@@ -60,23 +61,26 @@ exports.getAllProducts = async (req, res, next) => {
         const discountQuery = discount ? {discount : {$gte : +discount[0],$lte : +discount[1]}} : placeHolderQ
         let categorySearchQuery = placeHolderQ
         let brandQuery = placeHolderQ
-        if(brand) {
 
-            console.log(req.query)
+        
+        if(brand) {
             const singleBrand = await Brand.findOne({title : brand})
-            console.log(singleBrand)
             brandQuery =  {brand : {_id : singleBrand.id}} 
         }
         if (category) {
-            const categoryQuery = req.query.categories.split(',')
-            const categories = []
+            let categoryQuery = _.concat(req.query.categories)
+            let categories = []
+            
             for (c of categoryQuery) {
                 const category = await Category.findOne({ title: c })
-                categories.push({ _id: category.id })
+                if(category) {
+                    categories.push({ _id: category.id })
+                }
             }
+
             categorySearchQuery = { categories: { $all: categories } }
         }
-                            //sort
+        //sort
         let sortByQuery = {}
         if(sortBy) {
             switch (sortBy) {
@@ -86,12 +90,12 @@ exports.getAllProducts = async (req, res, next) => {
                 case "latest":
                     sortByQuery = {createdAt: "desc"}
                     break;
-                case "costly":
+                    case "costly":
                     sortByQuery = {price: "desc"}
                     break;
-                case "cheap":
-                    sortByQuery = {price: "asc"}
-                    break;
+                    case "cheap":
+                        sortByQuery = {price: "asc"}
+                        break;
                 default:
                     sortByQuery = {createdAt: "desc"}
                     break;
@@ -101,7 +105,7 @@ exports.getAllProducts = async (req, res, next) => {
         }
 
            //send Products
-            const sendData = (products, numberOfItems) => {
+           const sendData = (products, numberOfItems) => {
             if (!products || products.length == 0) {
                 res.status(404).json({ "message": "محصولی وجود ندارد", products: [] })
             } else {
@@ -113,18 +117,20 @@ exports.getAllProducts = async (req, res, next) => {
                 })
             }
         }
-
+        
         if (searching) {
             const numberOfFilteredItems = await Product.find({ ...textSearchQuery, ...categorySearchQuery,...brandQuery,...priceQuery,...discountQuery  }).countDocuments()
-            const filteredproducts = await Product.find({ ...textSearchQuery, ...categorySearchQuery ,...brandQuery,...priceQuery,...discountQuery}).sort({
+            const filteredproducts = await Product.find({...textSearchQuery, ...categorySearchQuery ,...brandQuery,...priceQuery,...discountQuery}).sort({
                 ...sortByQuery
             }).populate('categories').populate('brand').limit(itemPerPage).skip((page - 1) * itemPerPage)
+            // ...textSearchQuery, ...categorySearchQuery ,...brandQuery,...priceQuery,...discountQuery
             sendData(filteredproducts, numberOfFilteredItems)
         } else {
             const numberOfItems = await Product.find().countDocuments()
             const products = await Product.find().sort({
                 ...sortByQuery
             }).populate('categories').populate('brand').limit(itemPerPage).skip((page - 1) * itemPerPage)
+
             sendData(products, numberOfItems)
         }
