+++ server/models/validation/productValidation.js
@@ -2,7 +2,7 @@ const Yup = require('yup')
 
 exports.schema = Yup.object().shape({
     name : Yup.string().required("لطفا نام محصول را وارد کنید."),
-    price : Yup.string().required("لطفا قیمت مورد نظر را وارد کنید"),
+    price : Yup.number().required("لطفا قیمت مورد نظر را وارد کنید"),
     thumbnail : Yup.object().shape({
         size : Yup.number().max(3000000,'شما نمیتوانید بیشتر از 3 مگابایت اپلود کنید.'),
         name : Yup.string().required('لطا عکس بند انگشتی را انتخاب کنید.'),
