+++ server/package.json
@@ -25,6 +25,7 @@
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.21",
     "mongoose": "^6.4.6",
+    "mongoose-long": "^0.6.0",
     "nodemon": "^2.0.19",
     "short-uuid": "^4.2.0",
     "yup": "^0.32.11"
